requirements.txt
pycryptodome
python 3.9.1